VsInputComponent now implements the ControlValueAccessor interface. This interface is required for custom form controls to work with Angular forms.
Added NG_VALUE_ACCESSOR provider to the VsInputComponent to allow Angular to recognize it as a custom form control.

Added getter and setter for value with associated onChange and onTouch methods in the VsInputComponent.

Updated the template of VsInputComponent to use two-way data binding with ngModel for the value of the input field.

These changes allow VsInputComponent to be used as a form control in any Angular form.